Extract fakeInformer to shared internal testutil package
Both metriccontroller and pacemaker tests had identical 47-line
fakeInformer implementations. Move to pkg/tnf/internal/testutil/
so future interface changes only need one update.
